lab 16 (1) fix

+++ lab 16 (2).cpp
@@ -3,13 +3,15 @@
 using namespace std;
 
 bool isSorted(const vector<int> arr) {
-    for (size_t i = 0; i < arr.size() - 1; i++) {
+    for (int i = 0; i < arr.size() - 1; i++) {
         if (arr[i] > arr[i + 1]) {
             return false;
-        } 
+        }
     }
-    return true; 
+
+    return true;
 }
+
 
 int main() {
